requirements.txt
asgiref==3.2.7
boto3==1.14.26
botocore==1.17.26
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
dj-rest-auth==1.1.0
Django==3.0.8
django-allauth==0.42.0
django-cors-headers==3.4.0
django-extensions==3.0.3
django-filter==2.2.0
django-heroku==0.3.1
django-storages==1.9.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
docutils==0.15.2
factory-boy==2.12.0
Faker==4.1.1
gunicorn==20.0.4
idna==2.10
jmespath==0.10.0
mysqlclient==1.4.6
oauthlib==3.1.0
Pillow==7.2.0
psycopg2==2.8.5
psycopg2-binary==2.8.4
PyJWT==1.7.1
python-dateutil==2.8.1
python-decouple==3.3
python3-openid==3.2.0
pytz==2019.3
requests==2.24.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
urllib3==1.25.9
whitenoise==5.1.0
